description
This file contains the codes which are mainly used for the automation and testing of the Facebook login page. This login page contains the form where the user can input username/email and password. Moreover, users can create new accounts and new pages as well. We automate the login page by passing username and password keys.Finally, we can test whether the functions of login page are working or not.